# packages/dgcpos-client/openapi/community.yaml
openapi: 3.1.0
info:
  title: DGCPOS Community API
  version: 0.2.0
  description: |
    Core REST API for DGCPOS Community Edition (self-hosted POS).
    Enterprise-only paths (hospitality, live payments, admin) return 403 in CE.
  license:
    name: CC-BY-4.0
servers:
  - url: http://localhost:5000/api
    description: Local development
  - url: https://api.dgcpos.net/api
    description: Hosted SaaS (Enterprise by default)
security:
  - bearerAuth: []
...
paths:
...
  /sales/{id}:
    get:
      tags: [Sales]
      summary: Get sale by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Sale
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT